--- packages/builder/src/settings/pages/connections/knowledgebase/_components/KnowledgeConnectionReconnectRenderer.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-  import { ActionButton } from "@budibase/bbui"
-
-  interface KnowledgeConnectionRow {
-    id: string
-    reconnect?: () => void
-  }
-
-  export let row: KnowledgeConnectionRow
-</script>
-
-<ActionButton size="S" on:click={() => row.reconnect?.()}>
-  Reconnect
-</ActionButton>
